server/requirements.txt
asgiref==3.5.2
Django==3.2.14
django-cors-headers==3.13.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
mysqlclient==2.1.1
Pillow==9.2.0
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
